1.0.3
Okay, so for some reason, github was giving me problems when uplaoding the files before, and not uploading everything, so from now on I am just going to zip everything up and post it here like normal. When you download it, just unzip the file and use like normal. Sorry for the problems before, been trying to get it fixed. Keep on testing!
Thanks.